Decrease bundle size by removing the exclusion of `node_modules`
This allows the third party code to go through the same loader chain with `build-optimizer`, which results in the decrease of `vendor.js` bundle size from 515kb to 390kb.
This doesn't work when building the `server` bundles (strange error, see the comment in code), so the exclusion of `node_modules` is conditional.